fix(ffi): audit and fix remaining size_t/ssize_t width mismatches
Signatures confirmed against installed headers (glibc, libusb-1.0)
and, for libfreefare's already-documented comment, upstream
freefare.h:

- libc: calloc(size_t, size_t) and poll's nfds_t (unsigned long) were
  bound as int32/uint32; widen to uint. CAlloc()/Poll() callers
  updated accordingly.
- gousb: libusb_get_device_list returns ssize_t, not int; widen the
  binding to int and adjust the one caller (errorFromRet still takes
  the narrower int32 error-code type, so convert at the call site).
- libnfc: nfc_list_devices returns size_t and takes a size_t length
  (was uint32/uint32); nfc_initiator_select_passive_target,
  nfc_initiator_list_passive_targets and nfc_initiator_poll_target
  all take a size_t length/count (was int32). These are currently
  unreferenced bindings, so no other call sites to update.
- libfreefare: mifare_application_read is `ssize_t (..., size_t
  nbytes, ...)` per its own doc comment and upstream freefare.h; was
  bound as (..., uint32, ...) int32. Unreferenced binding, widened
  in place.
- libudev: udev_device_new_from_devnum takes a dev_t, which glibc
  fixes at 64 bits (__UQUAD_TYPE) on every platform, not merely
  pointer-width; was bound as uint32, widened to uint64. Unreferenced
  binding.

Audited but left unchanged (confirmed correct C `int`/`unsigned int`,
not size_t): plutobook (plutobook.h has no size_t in its public API,
only int/unsigned int - plutobook_load_html's `length` is `int`),
libusb interface/alternate-setting/endpoint params and
libusb_free_device_list's unref_devices (all plain `int`), sane's
SANE_Word/SANE_Bool/etc (intentionally int32 per the SANE spec),
libevdev (its only size_t-taking symbols, the *_from_name_n helpers,
aren't bound here), fontconfig and vlc's single int-returning calls.

Not auditable on this host (no libcec headers/pkg-config available)
and left untouched: cec/. Its bound functions take no size_t-shaped
parameters as far as the existing (mostly parameterless stub)
declarations go, but this wasn't checked against a real header.

+++ gousb/context.go
@@ -149,7 +149,7 @@ func (c *Context) OpenDevices(opener func(desc *DeviceDesc) bool) ([]*Device, er
 	var lst *libusbDevice
 	size := libusbGetDeviceList(c.ptr, &lst)
 	if size < 0 {
-		err := errorFromRet(size)
+		err := errorFromRet(int32(size))
 		return nil, err
 	}
 	defer libusbFreeDeviceList(lst, 0)
